Fixed a few more issues
