+++ core_types/src/cas/content.rs
@@ -4,7 +4,7 @@
 //! A test suite for AddressableContent is also implemented here.
 
 use crate::{
-    cas::storage::ContentAddressableStorage, error::error::HolochainError, hash::HashString,
+    cas::storage::ContentAddressableStorage, error::HolochainError, hash::HashString,
     json::JsonString,
 };
 use multihash::Hash;
